Refactor : Merge Dashboard Logic into a Single Module
